Get modified values from edit form
